fpga: zynqmp: Make word align the configuration data
To avoid the unwanted copies at firmware(PMUFW) this patch provides
a fix to align PL configuration data if the data is not word-aligned.
It improves the PL configuration timings to process unaligned bit
files.

To align the configuration data this patch adds a few padding bytes and
these additional padding bytes will not create any functional impact on
the PL configuration.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
Link: https://lore.kernel.org/all/20230314094222.66916-1-nava.kishore.manne@amd.com/
State: not-upstreamable
Reason: Framework refactoring is needed to address multiple-copies/alignment
issues.