mtd: spi-nor: Ensure 4-byte mode is set for all Micron multi-die flashes in stacked configuration
When Micron multi-die flashes are connected in stacked mode, the late_init
currently configures only the lower flash for 4-byte addressing. This leads
to a mismatch between the two flashes, causing RX tuning failures in DDR
mode for the upper flash. Updates the late_init logic to ensure that all
connected Micron multi-die flashes are correctly set to 4-byte mode.
Additionally, the resume logic is modified to reconfigure all connected
flashes to 4-byte mode if the addr_mode_byte flag is set.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Message-ID: <20251111154226.2729000-1-amit.kumar-mahapatra@amd.com>