ASoC: xlnx: Remove duplicate XLNX_AUD_FS_MULTIPLIER setting in hw_params
The hw_params() function correctly calculates the mclk_fs and writes
it to the XLNX_AUD_FS_MULTIPLIER register at the start of the function.
When used with the pl_snd_card, this does not impact functionality.

However, when used with either the simple-sound-card or audio-graph-card
drivers, the function incorrectly casts the private data of the selected
card driver to the pl_card_data type.
This results in writing an invalid value to the register.

Signed-off-by: Bram Vlerick <bram.vlerick@openpixelsystems.org>
Signed-off-by: Dhanunjanrao Katta <katta.dhanunjanrao@amd.com>
Message-ID: <1765361730-2855591-1-git-send-email-katta.dhanunjanrao@amd.com>